add ensure_no_static to tload/tstore
also move the logic functions to cancun/logic.py
