+++ eth/vm/logic/storage.py
@@ -144,20 +144,3 @@ def net_sstore(gas_schedule: NetSStoreGasSchedule, computation: ComputationAPI) 
         value=value,
     )
     return slot
-
-
-# -- transient storage -- #
-
-
-def tstore(computation: ComputationAPI) -> None:
-    address = computation.msg.storage_address
-    slot = computation.stack_pop1_int()
-    value = computation.stack_pop1_bytes()
-    computation.state.set_transient_storage(address, slot, value)
-
-
-def tload(computation: ComputationAPI) -> None:
-    address = computation.msg.storage_address
-    slot = computation.stack_pop1_int()
-    value = computation.state.get_transient_storage(address, slot)
-    computation.stack_push_bytes(value)
